Darkhorse (#2)
* DOIT

* temp checkin

* new approach. New structures. Thanks dxnn.

* ...

* mess mess

* super buggy sketch

* less buggy now

* new shapes, more refactoring!

* Clean up old code + rename new preposition functions

* Move funcs around

* Switch to 0.7 sweet js for access to infix.

* Infix macros working now

note that there was some hackiness with what SJS is capabale of handling
in terms of infix macros and bordering parens. B/c parens are delimiters for
the reader, it causes issue; so I had to change some of the code so that
the function reference is passed rather than the result of calling it.

* renaming files and cleaning up folder structure.

* Add better build process

* Modularize scripts (dsl vs library)